+++ api/effects-api/src/main/java/io/github/dailystruggle/effectsapi/fabric/FabricHandles.java
@@ -312,7 +312,7 @@ public final class FabricHandles implements HandleProvider {
                 final UUID id = player.getUUID();
                 net.minecraft.server.MinecraftServer server = player.getServer();
                 if (server != null) {
-                    final long deadline = server.getTickCount() + landingTimeoutTicks;
+                    final long deadline = (long) server.getTickCount() + landingTimeoutTicks;
                     final Runnable[] watchdog = new Runnable[1];
                     watchdog[0] = () -> {
                         if (server.getTickCount() >= deadline) {
