+++ platforms/rtp-fabric/rtp-fabric-v26_2_R1/src/main/java/io/github/dailystruggle/rtp/fabric/v26_2_R1/V26_2_R1FabricVersionAdapter.java
@@ -690,7 +690,7 @@ public final class V26_2_R1FabricVersionAdapter implements FabricVersionAdapter 
             int rr = (rgb >> 16) & 0xFF;
             int gg = (rgb >> 8) & 0xFF;
             int bb = rgb & 0xFF;
-            long dr = r - rr, dg = g - gg, db = b - bb;
+            long dr = (long) r - rr, dg = (long) g - gg, db = (long) b - bb;
             long dist = dr * dr + dg * dg + db * db;
             if (dist < bestDist) {
                 bestDist = dist;
