+++ rtp-core/src/main/java/io/github/dailystruggle/rtp/common/selection/region/selectors/memory/shapes/Circle.java
@@ -127,7 +127,7 @@ public class Circle extends MemoryShape<GenericMemoryShapeParams> {
       rotation += 0.25;
     }
 
-    double radius = ((long) (Math.sqrt(x * x + z * z)));
+    double radius = ((long) (Math.sqrt((double) x * x + (double) z * z)));
 
     return (long) ((radius * radius - cr * cr) * Math.PI + rotation * (2 * radius * Math.PI));
   }
@@ -151,7 +151,7 @@ public class Circle extends MemoryShape<GenericMemoryShapeParams> {
       rotation += 0.25;
     }
 
-    double radius = ((long) (Math.sqrt(x * x + z * z)));
+    double radius = ((long) (Math.sqrt((double) x * x + (double) z * z)));
 
     return (long) ((radius * radius - cr * cr) * Math.PI + rotation * (2 * radius * Math.PI));
   }
