+++ rtp-core/src/main/java/io/github/dailystruggle/rtp/common/selection/region/selectors/memory/shapes/Ellipse.java
@@ -156,7 +156,7 @@ public class Ellipse extends MemoryShape<EllipseMemoryShapeParams> {
       rotation += 0.25;
     }
 
-    double radius = ((long) (Math.sqrt(x * x + z * z)));
+    double radius = ((long) (Math.sqrt((double) x * x + (double) z * z)));
 
     return (long) ((radius * radius - cr * cr) * Math.PI + rotation * (2 * radius * Math.PI));
   }
